+++ src/components/index.js
@@ -164,11 +164,6 @@ export { EuiLink } from './link';
 
 export { EuiListGroup, EuiListGroupItem } from './list_group';
 
-// export { EuiStep, EuiSteps, EuiSubSteps, EuiStepsHorizontal } from './steps';
-
-export { EuiSuggestInput, EuiSuggestItem, EuiSuggest } from './suggest';
-// export { EuiSuggestItem } from './suggest_item';
-
 export {
   EUI_MODAL_CANCEL_BUTTON,
   EUI_MODAL_CONFIRM_BUTTON,
@@ -238,6 +233,8 @@ export { EuiSpacer } from './spacer';
 export { EuiStat } from './stat';
 
 export { EuiStep, EuiSteps, EuiSubSteps, EuiStepsHorizontal } from './steps';
+
+export { EuiSuggestInput, EuiSuggestItem, EuiSuggest } from './suggest';
 
 export {
   EuiTable,